Fixed warnings

+++ MegaEngine/Core/SceneManager.cpp
@@ -211,7 +211,7 @@ std::vector<GameObject*> SceneManager::getGameObjectsByName(const std::string& n
 
 	for (size_t i = 0; i < attached.size(); i++)
 	{
-		for (size_t j = 1; j < counter + 1; j++)
+		for (Uint16 j = 1; j < counter + 1; j++)
 		{
 			if (attached[i]->getName() == name + std::to_string(j))
 			{
